fix metering creation logic to skip if it's already installed.

+++ features/step_definitions/metering.rb
@@ -203,7 +203,7 @@ Given /^I enable route for#{OPT_QUOTED} metering service$/ do | metering_name |
   unless route('metering').exists?
     org_user = user
     ### XXX: TODO until 4.2 is GAed,
-    if operator_group('metering_operators').exists?
+    if operator_group('metering-operators').exists?
       cb[:metering_resource_type] = 'meteringconfig'
     else
       cb[:metering_resource_type] = 'metering'
@@ -256,33 +256,45 @@ end
 Given /^the#{OPT_QUOTED} metering service is installed(?: to $QUOTED)? using OLM$/ do | metering_ns |
   ensure_admin_tagged
   ensure_destructive_tagged
-
   # 1. create the metering namespace
   metering_ns ||= "openshift-metering"
-  project(metering_ns)
   step %Q/I switch to cluster admin pseudo user/
-  @result = user.cli_exec(:create_namespace, name: metering_ns)
-  # prep for OLM need to define configs via YAML
-  # 1. metering-catalogsourceconfig.yaml
-  catalog_source_config = "https://raw.githubusercontent.com/operator-framework/operator-metering/master/manifests/deploy/openshift/olm/metering.catalogsourceconfig.yaml"
-  # 2. metering-operatorgroup.yaml
-  operator_group = "https://raw.githubusercontent.com/operator-framework/operator-metering/master/manifests/deploy/openshift/olm/metering.operatorgroup.yaml"
-  # 3. metering-subscription.yaml
-  subscription = "https://raw.githubusercontent.com/operator-framework/operator-metering/master/manifests/deploy/openshift/olm/metering.subscription.yaml"
-  # XXX: this will need to be updted after 4.2 is released to the wild
-  metering_crd = "https://raw.githubusercontent.com/operator-framework/operator-metering/release-4.1/manifests/metering-config/default.yaml"
+  project(metering_ns)
+  if env.version_le('4.1', user: user)
+    metering_crd = "https://raw.githubusercontent.com/operator-framework/operator-metering/release-4.1/manifests/metering-config/default.yaml"
+  else
+    metering_crd = "https://raw.githubusercontent.com/operator-framework/operator-metering/master/manifests/metering-config/default.yaml"
+  end
+  begin
+    metering_installed = metering('operator-metering').exists?
+    cb.metering_resource_type = "metering"
+  rescue
+    cb.metering_resource_type = "meteringconfig"
+    metering_installed = metering_config('operator-metering').exists?
+  end
 
-  # should really get the namespace dynamically
-  @result = user.cli_exec(:apply, f: catalog_source_config, n: 'openshift-marketplace')
-  @result = user.cli_exec(:apply, f: operator_group, n: project.name)
-  @result = user.cli_exec(:apply, f: subscription, n: project.name)
-  # wait for initial metering pod to become running first and then apply CRD
-  step %Q/a pod becomes ready with labels:/, table(%{
-    | app=metering-operator |
-  })
+  unless project(metering_ns).exists? and metering_installed
+    @result = user.cli_exec(:create_namespace, name: metering_ns)
+    # prep for OLM need to define configs via YAML
+    # 1. metering-catalogsourceconfig.yaml
+    catalog_source_config = "https://raw.githubusercontent.com/operator-framework/operator-metering/master/manifests/deploy/openshift/olm/metering.catalogsourceconfig.yaml"
+    # 2. metering-operatorgroup.yaml
+    operator_group = "https://raw.githubusercontent.com/operator-framework/operator-metering/master/manifests/deploy/openshift/olm/metering.operatorgroup.yaml"
+    # 3. metering-subscription.yaml
+    subscription = "https://raw.githubusercontent.com/operator-framework/operator-metering/master/manifests/deploy/openshift/olm/metering.subscription.yaml"
+    # should really get the namespace dynamically
+    @result = user.cli_exec(:apply, f: catalog_source_config, n: 'openshift-marketplace')
+    @result = user.cli_exec(:apply, f: operator_group, n: project.name)
+    @result = user.cli_exec(:apply, f: subscription, n: project.name)
+    # wait for initial metering pod to become running first and then apply CRD
+    step %Q/a pod becomes ready with labels:/, table(%{
+      | app=metering-operator |
+    })
 
-  @result = user.cli_exec(:apply, f: metering_crd, n: project.name)
-  raise "OLM install of metering failed" unless @result[:success]
+    @result = user.cli_exec(:apply, f: metering_crd, n: project.name)
+    raise "OLM install of metering failed" unless @result[:success]
+  end
+
 end
 
 # XXX: currently OLM uninstall is TBD, we uninstall by removing the namespace
